Go back to int like main

+++ GEOS_Shared/GEOS_Utilities.F90
@@ -546,7 +546,7 @@
        end if
 
        TI = (TI - TMINTBL)*DEGSUBS+1
-       IT = nint(TI)
+       IT = int(TI)
 
        if(URAMP==TMIX) then
           DQ    = ESTBLX(IT+1) - ESTBLX(IT)
@@ -764,7 +764,7 @@
          end if
 
          TT = (TI - TMINTBL)*DEGSUBS+1
-         IT = nint(TT)
+         IT = int(TT)
 
          if(URAMP==TMIX) then
             DQQ =  ESTBLX(IT+1) - ESTBLX(IT)
